+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/PauseMenuStatusChangeListener.java
@@ -1,0 +1,7 @@
+package de.mephisto.vpin.commons.fx.pausemenu;
+
+public interface PauseMenuStatusChangeListener {
+
+  void pauseMenuShow();
+  void pauseMenuHide();
+}