+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarTableDetailsController.java
@@ -418,7 +418,7 @@ public class TablesSidebarTableDetailsController implements Initializable {
       GameEmulatorRepresentation gameEmulator = client.getEmulatorService().getGameEmulator(game.getEmulatorId());
       tableFilesBox.setVisible(gameEmulator != null && gameEmulator.isVpxEmulator());
 
-      autoFillBtn.setVisible(client.getEmulatorService().isVpxGame(game) && Features.FIELDS_STANDARD);
+      autoFillBtn.setVisible(client.getEmulatorService().isVpxOrFpGame(game) && Features.FIELDS_STANDARD);
 
       dateAdded.setText(game.getDateAdded() == null ? "-" : DateTimeFormatter.ofLocalizedDateTime(FormatStyle.MEDIUM).format(game.getDateAdded()));
       emulatorLabel.setText(client.getEmulatorService().getGameEmulator(game.getEmulatorId()).getName());