+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/util/SystemUtil.java
@@ -117,6 +117,7 @@ public class SystemUtil {
           String path = folder.getAbsolutePath();
 
           String remotePath = resolveNetworkPath(publicUrl, path);
+          LOG.info("Resolved network path '{}', use publicUrl '{}' and path '{}'", remotePath, publicUrl, path);
           if (remotePath != null) {
             openFolderWithOS(remotePath);
           }